+++ src/types/plugins/replication-graphql.d.ts
@@ -1,3 +1,4 @@
+import { ClientOptions } from 'graphql-ws';
 import { RxReplicationWriteToMasterRow } from '../replication-protocol.ts';
 import { ById, MaybePromise } from '../util.ts';
 import {
@@ -27,6 +28,7 @@ export type RxGraphQLReplicationPushQueryBuilder = (
     rows: RxReplicationWriteToMasterRow<any>[]
 ) => RxGraphQLReplicationQueryBuilderResponse;
 
+export type RxGraphQLPullWSOptions = Omit<ClientOptions, 'url' | 'shouldRetry' | 'webSocketImpl' | 'connectionParams'>;
 
 export type RxGraphQLReplicationPullQueryBuilder<CheckpointType> = (
     latestPulledCheckpoint: CheckpointType | undefined,
@@ -46,6 +48,7 @@ export type GraphQLSyncPullOptions<RxDocType, CheckpointType> = Omit<
     dataPath?: string | string[];
     responseModifier?: RxGraphQLPullResponseModifier<RxDocType, CheckpointType>;
     includeWsHeaders?: boolean;
+    wsOptions?: RxGraphQLPullWSOptions;
 };
 
 export type RxGraphQLPullResponseModifier<RxDocType, CheckpointType> = (